+++ elfi/visualization/interactive.py
@@ -5,14 +5,16 @@ import logging
 import matplotlib.pyplot as plt
 import numpy as np
 import scipy.interpolate
-from IPython import display
 
 logger = logging.getLogger(__name__)
 
 
-def plot_sample(samples, nodes=None, n=-1, displays=None, **opts):
-    """
-    Experimental, only dims 1-2 supported.
+def plot_sample(samples, nodes=None, n=-1, displays=None, **options):
+    """Plot a scatter-plot of samples.
+
+    Notes
+    -----
+    - Experimental, only dims 1-2 supported.
 
     Parameters
     ----------
@@ -23,7 +25,7 @@ def plot_sample(samples, nodes=None, n=-1, displays=None, **opts):
     displays : IPython.display.HTML
 
     """
-    axes = _prepare_axes(opts)
+    axes = _prepare_axes(options)
     if samples is None:
         return
     nodes = nodes or sorted(samples.keys())[:2]
@@ -41,42 +43,43 @@ def plot_sample(samples, nodes=None, n=-1, displays=None, **opts):
         axes.set_ylabel(nodes[1])
         axes.scatter(samples[nodes[0]][:n], samples[nodes[1]][:n])
 
-    if opts.get('interactive'):
-        update_interactive(displays, opts)
+    if options.get('interactive'):
+        update_interactive(displays, options)
         plt.close()
 
 
-def get_axes(**opts):
+def get_axes(**options):
     """Get an Axes object from `options`, or create one if needed."""
-    if 'axes' in opts:
-        return opts['axes']
+    if 'axes' in options:
+        return options['axes']
     return plt.gca()
 
 
-def update_interactive(displays, opts):
+def update_interactive(displays, options):
     """Update the interactive plot."""
     displays = displays or []
-    if opts.get('interactive'):
+    if options.get('interactive'):
+        from IPython import display
         display.clear_output(wait=True)
         displays.insert(0, plt.gcf())
         display.display(*displays)
 
 
-def _prepare_axes(opts):
-    axes = get_axes(**opts)
-    ion = opts.get('interactive')
+def _prepare_axes(options):
+    axes = get_axes(**options)
+    ion = options.get('interactive')
 
     if ion:
         axes.clear()
-    if opts.get('xlim'):
-        axes.set_xlim(opts.get('xlim'))
-    if opts.get('ylim'):
-        axes.set_ylim(opts.get('ylim'))
+    if options.get('xlim'):
+        axes.set_xlim(options.get('xlim'))
+    if options.get('ylim'):
+        axes.set_ylim(options.get('ylim'))
 
     return axes
 
 
-def draw_contour(fn, bounds, params=None, points=None, title=None, **opts):
+def draw_contour(fn, bounds, params=None, points=None, title=None, **options):
     """Plot a contour of a function.
 
     Experimental, only 2D supported.
@@ -93,8 +96,8 @@ def draw_contour(fn, bounds, params=None, points=None, title=None, **opts):
 
     """
     # Preparing the contour plot settings.
-    if opts.get('axes'):
-        axes = opts['axes']
+    if options.get('axes'):
+        axes = options['axes']
         plt.sca(axes)
     x, y = np.meshgrid(np.linspace(*bounds[0]), np.linspace(*bounds[1]))
     z = fn(np.c_[x.reshape(-1), y.reshape(-1)])